add biokg

+++ subgraph_extraction/graph_sampler.py
@@ -79,7 +79,7 @@ def links2subgraphs(A, graphs, params, max_label_value=None):
     links_length = 0
     for split_name, split in graphs.items():
         links_length += (len(split['pos']) + len(split['neg'])) * 2
-    map_size = links_length * BYTES_PER_DATUM
+    map_size = int(links_length * BYTES_PER_DATUM)
 
     env = lmdb.open(params.db_path, map_size=map_size, max_dbs=6)
